% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit.model}
\alias{fit.model}
\title{Fit model using machine learning}
\usage{
fit.model(training_data)
}
\arguments{
\item{training_data}{A list of ineage tree}
}
\value{
ML model for each relationship level
}
\description{
Fit model using machine learning
}
\author{
Renata Retkute, \email{r.retkute@yahoo.com}
}
