% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_features_singl_mut.R
\name{get.features.singl.mut}
\alias{get.features.singl.mut}
\title{Get features based on singe mutation}
\usage{
get.features.singl.mut(tree)
}
\arguments{
\item{tree}{Lineage tree}
}
\value{
List containing feature matrices & matrix showing relationships between cells
}
\description{
Get features based on singe mutation
}
\author{
Renata Retkute, \email{r.retkute@yahoo.com}
}
