% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lineage_thr_singl_mut.R
\name{make.lineage.thr.singl.mut}
\alias{make.lineage.thr.singl.mut}
\title{Make a lineage tree based on a vector of thresholds and single mutation}
\usage{
make.lineage.thr.singl.mut(xx, ML.model, thrs)
}
\arguments{
\item{xx}{Barcode data}

\item{thrs}{A vector of thresholds}

\item{fit.model}{A list of fitted ML models for relationship between cells}
}
\value{
List containing feature matrices & matrix showing relationships between cells
}
\description{
Make a lineage tree based on a vector of thresholds and single mutation
}
\author{
Renata Retkute, \email{r.retkute@yahoo.com}
}
