% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_tree.R
\name{sim_tree}
\alias{sim_tree}
\title{Simulate lineage tree}
\usage{
sim_tree(targets, d, mu, notmut, mut, plt = FALSE, full.div = FALSE)
}
\arguments{
\item{targets}{Number of targets}

\item{d}{Number of divisions}

\item{mu}{Mutation rate}

\item{notmut}{Symbol for not-mutated state}

\item{mut}{Symbol for mutated state}

\item{plt}{Plot tree at each division (default =FALSE)}

\item{full.div}{Return all division/mutation history (default =FALSE)}
}
\value{
Lineage tree, with id's containing mutation records
}
\description{
Simulate lineage tree
}
\author{
Renata Retkute, \email{r.retkute@yahoo.com}
}
